# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Digital Twin Fisiologico API",
    "description": "Stateless physiology analytics API. HTTP routers are thin; application services orchestrate engines under `engines/`.",
    "version": "5.1.0",
    "x-codegen": {
      "typescript_output": "frontend/src/api/generated/schema.ts",
      "regenerate": "make openapi-frontend"
    }
  },
  "paths": {
...
    "/twin/state/build": {
      "post": {
        "tags": [
          "twin"
        ],
        "summary": "Build TwinState v1",
        "description": "Create canonical twin_state.v1 blob for DB persistence.",
        "operationId": "twinStateBuild",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwinStateBuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Engine JSON payload — see docs/FRONTEND_DEVELOPER_GUIDE.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnginePayload",
                  "properties": {},
                  "additionalProperties": true,
                  "type": "object",
                  "title": "EnginePayload",
                  "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/twin/state/update-from-ride": {
      "post": {
        "tags": [
          "twin"
        ],
        "summary": "Update TwinState after ride",
        "operationId": "twinStateUpdateFromRide",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwinStateUpdateRideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Engine JSON payload — see docs/FRONTEND_DEVELOPER_GUIDE.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnginePayload",
                  "properties": {},
                  "additionalProperties": true,
                  "type": "object",
                  "title": "EnginePayload",
                  "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/twin/state/update-from-workout-result": {
      "post": {
        "tags": [
          "twin"
        ],
        "summary": "Append workout compliance to TwinState",
        "operationId": "twinStateUpdateFromWorkout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwinStateUpdateWorkoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Engine JSON payload — see docs/FRONTEND_DEVELOPER_GUIDE.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnginePayload",
                  "properties": {},
                  "additionalProperties": true,
                  "type": "object",
                  "title": "EnginePayload",
                  "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/twin/state/project": {
      "post": {
        "tags": [
          "twin"
        ],
        "summary": "Season what-if projection",
        "description": "Project CP, load and readiness from TwinState + future calendar.",
        "operationId": "twinStateProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeasonProjectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Engine JSON payload — see docs/FRONTEND_DEVELOPER_GUIDE.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnginePayload",
                  "properties": {},
                  "additionalProperties": true,
                  "type": "object",
                  "title": "EnginePayload",
                  "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projection/season": {
      "post": {
        "tags": [
          "twin"
        ],
        "summary": "Season projection (alias)",
        "description": "Alias of POST /twin/state/project.",
        "operationId": "projectionSeason",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeasonProjectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Engine JSON payload — see docs/FRONTEND_DEVELOPER_GUIDE.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnginePayload",
                  "properties": {},
                  "additionalProperties": true,
                  "type": "object",
                  "title": "EnginePayload",
                  "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnginePayload": {
        "properties": {},
        "additionalProperties": true,
        "type": "object",
        "title": "EnginePayload",
        "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SeasonProjectionRequest": {
        "properties": {
          "twin_state": {
            "additionalProperties": true,
            "type": "object",
            "title": "Twin State"
          },
          "calendar_plan": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Calendar Plan",
            "description": "Future planned workouts/assignments."
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date",
            "description": "ISO date projection start."
          },
          "target_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date",
            "description": "ISO date projection end."
          },
          "max_days": {
            "type": "integer",
            "maximum": 400.0,
            "minimum": 1.0,
            "title": "Max Days",
            "default": 365
          }
        },
        "type": "object",
        "required": [
          "twin_state"
        ],
        "title": "SeasonProjectionRequest"
      },
...
      "TwinStateBuildRequest": {
        "properties": {
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload",
            "description": "Initial athlete anchor, snapshot, curve and profile fragments."
          }
        },
        "type": "object",
        "title": "TwinStateBuildRequest"
      },
      "TwinStateUpdateRideRequest": {
        "properties": {
          "twin_state": {
            "additionalProperties": true,
            "type": "object",
            "title": "Twin State",
            "description": "Current persisted TwinState (twin_state.v1)."
          },
          "ride_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ride Summary",
            "description": "Output of POST /ride/summary."
          },
          "ingest_result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingest Result",
            "description": "Output of POST /ride/ingest."
          },
          "power_source_report": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power Source Report"
          },
          "ride_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ride Id"
          }
        },
        "type": "object",
        "required": [
          "twin_state"
        ],
        "title": "TwinStateUpdateRideRequest"
      },
      "TwinStateUpdateWorkoutRequest": {
        "properties": {
          "twin_state": {
            "additionalProperties": true,
            "type": "object",
            "title": "Twin State"
          },
          "compliance_result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Compliance Result",
            "description": "Output of POST /workouts/compare."
          },
          "assignment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignment Id"
          }
        },
        "type": "object",
        "required": [
          "twin_state",
          "compliance_result"
        ],
        "title": "TwinStateUpdateWorkoutRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "{baseUrl}",
      "description": "Configurable deployment",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:8000",
          "description": "Set via VITE_API_BASE_URL in the frontend"
        }
      }
    }
  ],
...
}